+++ apps/explorer/lib/explorer/chain/import/runner/logs.ex
@@ -6,7 +6,7 @@ defmodule Explorer.Chain.Import.Runner.Logs do
   require Ecto.Query
 
   alias Ecto.{Changeset, Multi, Repo}
-  alias Explorer.Chain.{Import, Log}
+  alias Explorer.Chain.{Import, Log, LogFirstTopic}
   alias Explorer.Prometheus.Instrumenter
 
   import Ecto.Query, only: [from: 2]
@@ -64,8 +64,41 @@ defmodule Explorer.Chain.Import.Runner.Logs do
   defp insert(repo, changes_list, %{timeout: timeout, timestamps: timestamps} = options) when is_list(changes_list) do
     on_conflict = Map.get_lazy(options, :on_conflict, &default_on_conflict/0)
 
+    log_first_topic_changes_list =
+      changes_list
+      |> Enum.filter(&(Map.has_key?(&1, :first_topic) && &1.first_topic))
+      |> Enum.map(&%{:hash => &1.first_topic})
+      |> Enum.uniq()
+
+    {:ok, _} =
+      Import.insert_changes_list(
+        repo,
+        log_first_topic_changes_list,
+        conflict_target: [:hash],
+        on_conflict: :nothing,
+        for: LogFirstTopic,
+        returning: true,
+        timeout: timeout,
+        timestamps: timestamps
+      )
+
+    log_first_topic_hashes =
+      log_first_topic_changes_list
+      |> Enum.map(& &1.hash)
+      |> Enum.uniq()
+
+    log_first_topic_ids_hashes = LogFirstTopic.get_log_first_topic_ids_by_hashes(log_first_topic_hashes)
+
+    changes_list_with_log_first_topic_id =
+      if Enum.count(log_first_topic_ids_hashes) > 0 do
+        put_log_first_topic_id(changes_list, log_first_topic_ids_hashes)
+      else
+        changes_list
+      end
+
     # Enforce Log ShareLocks order (see docs: sharelocks.md)
-    ordered_changes_list = Enum.sort_by(changes_list, &{&1.transaction_hash, &1.block_hash, &1.index})
+    ordered_changes_list =
+      Enum.sort_by(changes_list_with_log_first_topic_id, &{&1.transaction_hash, &1.block_hash, &1.index})
 
     {:ok, _} =
       Import.insert_changes_list(
@@ -80,6 +113,19 @@ defmodule Explorer.Chain.Import.Runner.Logs do
       )
   end
 
+  defp put_log_first_topic_id(changes_list, log_first_topic_ids_hashes) do
+    Enum.map(changes_list, fn item ->
+      {first_topic_id, _} =
+        Enum.find(log_first_topic_ids_hashes, fn {_, hash} ->
+          hash == item.first_topic
+        end)
+
+      item
+      |> Map.put(:first_topic_id, first_topic_id)
+      |> Map.drop([:first_topic])
+    end)
+  end
+
   defp default_on_conflict do
     from(
       log in Log,
@@ -87,7 +133,7 @@ defmodule Explorer.Chain.Import.Runner.Logs do
         set: [
           address_hash: fragment("EXCLUDED.address_hash"),
           data: fragment("EXCLUDED.data"),
-          first_topic: fragment("EXCLUDED.first_topic"),
+          first_topic_id: fragment("EXCLUDED.first_topic_id"),
           second_topic: fragment("EXCLUDED.second_topic"),
           third_topic: fragment("EXCLUDED.third_topic"),
           fourth_topic: fragment("EXCLUDED.fourth_topic"),
@@ -99,10 +145,10 @@ defmodule Explorer.Chain.Import.Runner.Logs do
       ],
       where:
         fragment(
-          "(EXCLUDED.address_hash, EXCLUDED.data, EXCLUDED.first_topic, EXCLUDED.second_topic, EXCLUDED.third_topic, EXCLUDED.fourth_topic) IS DISTINCT FROM (?, ?, ?, ?, ?, ?)",
+          "(EXCLUDED.address_hash, EXCLUDED.data, EXCLUDED.first_topic_id, EXCLUDED.second_topic, EXCLUDED.third_topic, EXCLUDED.fourth_topic) IS DISTINCT FROM (?, ?, ?, ?, ?, ?)",
           log.address_hash,
           log.data,
-          log.first_topic,
+          log.first_topic_id,
           log.second_topic,
           log.third_topic,
           log.fourth_topic