+++ apps/explorer/lib/explorer/chain/log.ex
@@ -6,13 +6,13 @@ defmodule Explorer.Chain.Log do
   require Logger
 
   alias ABI.{Event, FunctionSelector}
-  alias Explorer.Chain
-  alias Explorer.Chain.{Address, Block, ContractMethod, Data, Hash, Log, Transaction}
+  alias Explorer.{Chain, Repo}
+  alias Explorer.Chain.{Address, Block, ContractMethod, Data, Hash, Log, LogFirstTopic, Transaction}
   alias Explorer.Chain.SmartContract.Proxy
   alias Explorer.SmartContract.SigProviderInterface
 
   @required_attrs ~w(address_hash data block_hash index transaction_hash)a
-  @optional_attrs ~w(first_topic second_topic third_topic fourth_topic block_number)a
+  @optional_attrs ~w(first_topic_id first_topic second_topic third_topic fourth_topic block_number)a
 
   @typedoc """
    * `address` - address of contract that generate the event
@@ -20,7 +20,7 @@ defmodule Explorer.Chain.Log do
    * `block_number` - The block number that the transfer took place.
    * `address_hash` - foreign key for `address`
    * `data` - non-indexed log parameters.
-   * `first_topic` - `topics[0]`
+   * `first_topic_id` - id of LogFirstTopic
    * `second_topic` - `topics[1]`
    * `third_topic` - `topics[2]`
    * `fourth_topic` - `topics[3]`
@@ -34,10 +34,11 @@ defmodule Explorer.Chain.Log do
           block_hash: Hash.Full.t(),
           block_number: non_neg_integer() | nil,
           data: Data.t(),
-          first_topic: Hash.Full.t(),
-          second_topic: Hash.Full.t(),
-          third_topic: Hash.Full.t(),
-          fourth_topic: Hash.Full.t(),
+          log_first_topic: %Ecto.Association.NotLoaded{} | Hash.Full.t() | nil,
+          first_topic_id: non_neg_integer() | nil,
+          second_topic: Hash.Full.t() | nil,
+          third_topic: Hash.Full.t() | nil,
+          fourth_topic: Hash.Full.t() | nil,
           transaction: %Ecto.Association.NotLoaded{} | Transaction.t(),
           transaction_hash: Hash.Full.t(),
           index: non_neg_integer()
@@ -46,7 +47,7 @@ defmodule Explorer.Chain.Log do
   @primary_key false
   schema "logs" do
     field(:data, Data)
-    field(:first_topic, Hash.Full)
+    field(:first_topic, Hash.Full, virtual: true)
     field(:second_topic, Hash.Full)
     field(:third_topic, Hash.Full)
     field(:fourth_topic, Hash.Full)
@@ -70,6 +71,13 @@ defmodule Explorer.Chain.Log do
       references: :hash,
       type: Hash.Full
     )
+
+    belongs_to(:log_first_topic, LogFirstTopic,
+      foreign_key: :first_topic_id,
+      primary_key: true,
+      references: :id,
+      type: :integer
+    )
   end
 
   @doc """
@@ -81,12 +89,12 @@ defmodule Explorer.Chain.Log do
       ...>     address_hash: "0x8bf38d4764929064f2d4d3a56520a76ab3df415b",
       ...>     block_hash: "0xf6b4b8c88df3ebd252ec476328334dc026cf66606a84fb769b3d3cbccc8471bd",
       ...>     data: "0x000000000000000000000000862d67cb0773ee3f8ce7ea89b328ffea861ab3ef",
-      ...>     first_topic: "0x600bcf04a13e752d1e3670a5a9f1c21177ca2a93c6f5391d4f1298d098097c22",
       ...>     fourth_topic: nil,
       ...>     index: 0,
       ...>     second_topic: nil,
       ...>     third_topic: nil,
-      ...>     transaction_hash: "0x53bd884872de3e488692881baeec262e7b95234d3965248c39fe992fffd433e5"
+      ...>     transaction_hash: "0x53bd884872de3e488692881baeec262e7b95234d3965248c39fe992fffd433e5",
+      ...>     first_topic_id: 1
       ...>   }
       ...> )
       iex> changeset.valid?
@@ -183,10 +191,10 @@ defmodule Explorer.Chain.Log do
   end
 
   defp find_method_candidates(log, transaction, options, events_acc, skip_sig_provider?) do
-    if is_nil(log.first_topic) do
+    if is_nil(log.log_first_topic) do
       {{:error, :could_not_decode}, events_acc}
     else
-      <<method_id::binary-size(4), _rest::binary>> = log.first_topic.bytes
+      <<method_id::binary-size(4), _rest::binary>> = log.log_first_topic.hash.bytes
 
       if Map.has_key?(events_acc, method_id) do
         {events_acc[method_id], events_acc}
@@ -235,7 +243,7 @@ defmodule Explorer.Chain.Log do
            abi
            |> ABI.parse_specification(include_events?: true)
            |> Event.find_and_decode(
-             log.first_topic && log.first_topic.bytes,
+             log.log_first_topic && log.log_first_topic.hash.bytes,
              log.second_topic && log.second_topic.bytes,
              log.third_topic && log.third_topic.bytes,
              log.fourth_topic && log.fourth_topic.bytes,
@@ -280,7 +288,7 @@ defmodule Explorer.Chain.Log do
          {:ok, result} <-
            SigProviderInterface.decode_event(
              [
-               log.first_topic,
+               log.log_first_topic.hash,
                log.second_topic,
                log.third_topic,
                log.fourth_topic
@@ -318,8 +326,11 @@ defmodule Explorer.Chain.Log do
 
   def fetch_log_by_tx_hash_and_first_topic(tx_hash, first_topic, options \\ []) do
     __MODULE__
-    |> where([l], l.transaction_hash == ^tx_hash and l.first_topic == ^first_topic)
+    |> join(:left, [l], topic in LogFirstTopic, on: l.first_topic_id == topic.id)
+    |> where([l, topic], l.transaction_hash == ^tx_hash and topic.hash == ^first_topic)
     |> limit(1)
+    |> select([l, topic], l)
     |> Chain.select_repo(options).one()
+    |> Repo.preload(:log_first_topic)
   end
 end