+++ apps/explorer/priv/repo/migrations/20220315093927_add_block_timestamp_to_transaction_table.exs
@@ -1,0 +1,11 @@
+defmodule Explorer.Repo.Migrations.AddBlockTimestampToTransactionTable do
+  use Ecto.Migration
+
+  def change do
+    alter table("transactions") do
+      add(:block_timestamp, :utc_datetime_usec)
+    end
+
+    create(index(:transactions, :block_timestamp))
+  end
+end