+++ apps/explorer/test/explorer/chain/import_test.exs
@@ -22,14 +22,12 @@ defmodule Explorer.Chain.ImportTest do
 
   @moduletag :capturelog
 
-  @first_topic_hex_string "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
   @second_topic_hex_string "0x000000000000000000000000e8ddc5c7a2d2f0d7a9798459c0104fdf5e987aca"
   @third_topic_hex_string "0x000000000000000000000000515c09c5bba1ed566b02a5b0599ec5d5d0aee73d"
 
   doctest Import
 
   describe "all/1" do
-    {:ok, first_topic} = Explorer.Chain.Hash.Full.cast(@first_topic_hex_string)
     {:ok, second_topic} = Explorer.Chain.Hash.Full.cast(@second_topic_hex_string)
     {:ok, third_topic} = Explorer.Chain.Hash.Full.cast(@third_topic_hex_string)
     # set :timeout options to cover lines that use the timeout override when available
@@ -98,7 +96,6 @@ defmodule Explorer.Chain.ImportTest do
             block_hash: "0xf6b4b8c88df3ebd252ec476328334dc026cf66606a84fb769b3d3cbccc8471bd",
             address_hash: "0x8bf38d4764929064f2d4d3a56520a76ab3df415b",
             data: "0x0000000000000000000000000000000000000000000000000de0b6b3a7640000",
-            first_topic: first_topic,
             second_topic: second_topic,
             third_topic: third_topic,
             fourth_topic: nil,
@@ -171,7 +168,6 @@ defmodule Explorer.Chain.ImportTest do
     }
 
     test "with valid data" do
-      {:ok, first_topic} = Explorer.Chain.Hash.Full.cast(@first_topic_hex_string)
       {:ok, second_topic} = Explorer.Chain.Hash.Full.cast(@second_topic_hex_string)
       {:ok, third_topic} = Explorer.Chain.Hash.Full.cast(@third_topic_hex_string)
       difficulty = Decimal.new(340_282_366_920_938_463_463_374_607_431_768_211_454)
@@ -285,7 +281,6 @@ defmodule Explorer.Chain.ImportTest do
                           167, 100, 0, 0>>
                     },
                     index: 0,
-                    first_topic: ^first_topic,
                     second_topic: ^second_topic,
                     third_topic: ^third_topic,
                     fourth_topic: nil,