+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 - [#9055](https://github.com/blockscout/blockscout/pull/9055) - Add ASC indices for logs, token transfers, transactions
 - [#9038](https://github.com/blockscout/blockscout/pull/9038) - Token type filling migrations
+- [#9037](https://github.com/blockscout/blockscout/pull/9037) - Log first topic to separate table
 - [#9009](https://github.com/blockscout/blockscout/pull/9009) - Index for block refetch_needed
 - [#9007](https://github.com/blockscout/blockscout/pull/9007) - Drop logs type index
 - [#9006](https://github.com/blockscout/blockscout/pull/9006) - Drop unused indexes on address_current_token_balances table